chore: use Long for model ids

+++ app/src/main/java/com/assignment/jakewharton/model/Actor.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
 data class Actor(
-    @Expose val id: Int,
+    @Expose val id: Long,
     @Expose val url: String,
     @Expose val type: String,
     @Expose @SerializedName("display_login") val displayLogin: String
