+++ app/src/main/java/com/assignment/jakewharton/model/Event.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 import java.util.*
 
 data class Event(
-    @Expose val id: Int,
+    @Expose val id: Long,
     @Expose val type: String,
     @Expose val actor: Actor,
     @Expose @SerializedName("created_at") val created: Date
